Revert "Update KubeFedCluster CRD"
This reverts commit 41482fd420313f99f6d986ebfc538203912cbc8a.
