--- tools.go
@@ -1,8 +1,0 @@
-// +build tools
-
-package tools
-
-import (
-	// Code generators built at runtime.
-	_ "sigs.k8s.io/controller-tools/cmd/controller-gen"
-)
